refactored classes ParseModel and ExperimentOperations into separate file in a separate folder "experimentops" for easier overview
tested locally
